pyproject.toml
[project]
name = "workout-tracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.18.4",
    "asyncio>=4.0.0",
    "asyncpg==0.31.0",
    "email-validator==2.3.0",
    "fastapi==0.129.0",
    "psycopg2-binary==2.9.11",
    "pydantic-settings>=2.12.0",
    "ruff>=0.15.1",
    "uvicorn>=0.40.0",
]
